+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,11 +18,11 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
-* Supported local user IsNfSv3Enabled and ExtendedGroups
+* Supported local user IsNfSv3Enabled and ExtendedGroup
     - `Set-AzStorageLocalUser`
 * Supported list local user with IncludeNFSv3
     - `Get-AzStorageLocalUser`
-* Supported EnableExtendedGroups when creating and updating a storage account 
+* Supported EnableExtendedGroup when creating and updating a storage account 
     - `New-AzStorageAccount`
     - `Set-AzStorageAccount`